Restructure layout and add UI section
