+++ web/src/components/Cells/Organization/EditOrganizationCell/EditOrganizationCell.tsx
@@ -4,7 +4,7 @@ import type { CellSuccessProps, CellFailureProps } from '@redwoodjs/web'
 import { useMutation } from '@redwoodjs/web'
 import { toast } from '@redwoodjs/web/toast'
 import { routes } from '@redwoodjs/router'
-import Loader from 'src/components/Elements/Loader'
+import Loader from 'src/ui/Loader'
 import PageTitle from 'src/components/Layout/PageTitle'
 import OrganizationForm from 'src/components/Forms/OrganizationForm'
