+++ web/src/components/Cells/Project/ProjectCell/ProjectCell.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import type { FindProjectQuery } from 'types/graphql'
 import type { CellSuccessProps, CellFailureProps } from '@redwoodjs/web'
 import { Link, navigate, routes } from '@redwoodjs/router'
-import Loader from 'src/components/Elements/Loader'
+import Loader from 'src/ui/Loader'
 import PageTitle from 'src/components/Layout/PageTitle'
 import Popup from 'src/components/Elements/Popup'
 import NewTask from 'src/components/NewComponents/NewTask'
