+++ web/src/components/Cells/Project/ProjectsCell/ProjectsCell.tsx
@@ -2,7 +2,7 @@ import type { ProjectsQuery } from 'types/graphql'
 import type { CellSuccessProps, CellFailureProps } from '@redwoodjs/web'
 import { Link, routes } from '@redwoodjs/router'
 import { ChevronRightIcon } from '@heroicons/react/outline'
-import Loader from 'src/components/Elements/Loader'
+import Loader from 'src/ui/Loader'
 
 export const QUERY = gql`
   query ProjectsQuery {
