+++ web/src/components/Cells/Ticket/TicketsCell/TicketsCell.tsx
@@ -1,7 +1,7 @@
 import type { TicketsQuery } from 'types/graphql'
 import type { CellSuccessProps, CellFailureProps } from '@redwoodjs/web'
 import { Link, routes } from '@redwoodjs/router'
-import Loader from 'src/components/Elements/Loader'
+import Loader from 'src/ui/Loader'
 
 export const QUERY = gql`
   query TicketsQuery {
