+++ web/src/layouts/DashboardLayout/Footer/Footer.stories.tsx
@@ -1,0 +1,7 @@
+import Footer from './Footer'
+
+export const generated = () => {
+  return <Footer />
+}
+
+export default { title: 'Components/Footer' }
