+++ web/src/layouts/DashboardLayout/MobileSidebar/MobileSidebar.stories.tsx
@@ -1,0 +1,7 @@
+import MobileSidebar from './MobileSidebar'
+
+export const generated = () => {
+  return <MobileSidebar />
+}
+
+export default { title: 'Components/MobileSidebar' }
