+++ web/src/layouts/DashboardLayout/MobileSidebar/MobileSidebar.tsx
@@ -1,0 +1,85 @@
+import { Dialog, Transition } from '@headlessui/react'
+import { Link, routes } from '@redwoodjs/router'
+import { XIcon } from '@heroicons/react/outline'
+import Navigation from '../Navigation'
+import Logo from 'src/lib/logo.svg'
+
+type MobileSidebarProps = {
+  isOpen: boolean
+  setIsOpen: React.Dispatch<React.SetStateAction<boolean>>
+}
+
+const MobileSidebar = (props: MobileSidebarProps) => {
+  const { isOpen, setIsOpen } = props
+
+  return (
+    <Transition.Root show={isOpen} as={React.Fragment}>
+      <Dialog
+        as="div"
+        className="fixed inset-0 z-[80] flex md:hidden"
+        onClose={setIsOpen}
+      >
+        <Transition.Child
+          as={React.Fragment}
+          enter="transition-opacity ease-linear duration-300"
+          enterFrom="opacity-0"
+          enterTo="opacity-100"
+          leave="transition-opacity ease-linear duration-300"
+          leaveFrom="opacity-100"
+          leaveTo="opacity-0"
+        >
+          <Dialog.Overlay className="fixed inset-0 bg-gray-600 bg-opacity-75" />
+        </Transition.Child>
+        <Transition.Child
+          as={React.Fragment}
+          enter="transition ease-in-out duration-300 transform"
+          enterFrom="-translate-x-full"
+          enterTo="translate-x-0"
+          leave="transition ease-in-out duration-300 transform"
+          leaveFrom="translate-x-0"
+          leaveTo="-translate-x-full"
+        >
+          <div className="relative flex w-full max-w-xs flex-1 flex-col bg-gray-800 pb-4">
+            <Transition.Child
+              as={React.Fragment}
+              enter="ease-in-out duration-300"
+              enterFrom="opacity-0"
+              enterTo="opacity-100"
+              leave="ease-in-out duration-300"
+              leaveFrom="opacity-100"
+              leaveTo="opacity-0"
+            >
+              <div className="absolute top-0 right-0 -mr-12 pt-2">
+                <button
+                  type="button"
+                  className="ml-1 flex h-10 w-10 items-center justify-center rounded-full focus:outline-none focus:ring focus:ring-inset focus:ring-white"
+                  onClick={() => setIsOpen(false)}
+                >
+                  <span className="sr-only">Close sidebar</span>
+                  <XIcon className="h-6 w-6 text-white" aria-hidden="true" />
+                </button>
+              </div>
+            </Transition.Child>
+            <Link
+              to={routes.home()}
+              className="group flex flex-shrink-0 items-center bg-gray-900 px-4 py-4 transition-colors duration-300 hover:bg-indigo-900"
+            >
+              <Logo className="h-8 w-auto fill-current text-indigo-500 transition-colors duration-300 group-hover:text-white" />
+              <h2 className="ml-3 text-xl font-bold text-white">Stablocks</h2>
+            </Link>
+            <div className="mt-5 h-0 flex-1 overflow-y-auto">
+              <nav className="space-y-1 px-2">
+                <Navigation />
+              </nav>
+            </div>
+          </div>
+        </Transition.Child>
+        <div className="w-14 flex-shrink-0" aria-hidden="true">
+          {/* Dummy element to force sidebar to shrink to fit close icon */}
+        </div>
+      </Dialog>
+    </Transition.Root>
+  )
+}
+
+export default MobileSidebar
