+++ web/src/layouts/DashboardLayout/Navbar/Navbar.stories.tsx
@@ -1,0 +1,7 @@
+import Navbar from './Navbar'
+
+export const generated = () => {
+  return <Navbar />
+}
+
+export default { title: 'Components/Navbar' }
