+++ web/src/layouts/DashboardLayout/Navbar/Navbar.test.tsx
@@ -1,0 +1,14 @@
+import { render } from '@redwoodjs/testing/web'
+
+import Navbar from './Navbar'
+
+//   Improve this test with help from the Redwood Testing Doc:
+//    https://redwoodjs.com/docs/testing#testing-components
+
+describe('Navbar', () => {
+  it('renders successfully', () => {
+    expect(() => {
+      render(<Navbar />)
+    }).not.toThrow()
+  })
+})
