+++ web/src/layouts/DashboardLayout/Navbar/SearchPopup/SearchPopup.stories.tsx
