+++ web/src/layouts/DashboardLayout/Navbar/SearchPopup/SearchPopup.test.tsx
