+++ web/src/layouts/DashboardLayout/Navigation/Navigation.stories.tsx
