+++ web/src/layouts/DashboardLayout/StaticSidebar/StaticSidebar.stories.tsx
@@ -1,0 +1,7 @@
+import StaticSidebar from './StaticSidebar'
+
+export const generated = () => {
+  return <StaticSidebar />
+}
+
+export default { title: 'Components/StaticSidebar' }
