--- web/src/pages/Essentials/Organization/EditOrganizationPage/EditOrganizationPage.tsx
@@ -1,7 +1,0 @@
-import EditOrganizationCell from 'src/components/Essentials/Organization/EditOrganizationCell'
-
-const EditOrganizationPage = () => {
-  return <EditOrganizationCell />
-}
-
-export default EditOrganizationPage
