--- web/src/pages/Essentials/Organization/OrganizationPage/OrganizationPage.stories.tsx
@@ -1,7 +1,0 @@
-import OrganizationPage from './OrganizationPage'
-
-export const generated = () => {
-  return <OrganizationPage />
-}
-
-export default { title: 'Pages/OrganizationPage' }
