--- web/src/pages/Essentials/Organization/OrganizationPage/OrganizationPage.test.tsx
@@ -1,11 +1,0 @@
-import { render } from '@redwoodjs/testing/web'
-
-import OrganizationPage from './OrganizationPage'
-
-describe('OrganizationPage', () => {
-  it('renders successfully', () => {
-    expect(() => {
-      render(<OrganizationPage />)
-    }).not.toThrow()
-  })
-})
