+++ web/src/utils/base.ts
@@ -1,0 +1,3 @@
+export function classNames(...classes: string[]) {
+  return classes.filter(Boolean).join(' ')
+}
